requirements.txt
requests
bs4
pygithub
